implemented explore page.  Set up the dynamic loading.  Made the timer code reusable and added an NFTCard component to also make the cards reusable.  Used the filter function of the API to display filtered results
