+++ src/components/explore/ExploreItems.jsx
@@ -1,78 +1,60 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import AuthorImage from "../../images/author_thumbnail.jpg";
-import nftImage from "../../images/nftImage.jpg";
+import React, { useState } from "react";
+import NFTCard from "../common/NFTCard";
 
-const ExploreItems = () => {
+const ExploreItems = ({ items, loading, filter, onFilterChange }) => {
+  const [displayedItems, setDisplayedItems] = useState(8);
+
+  const handleLoadMore = () => {
+    setDisplayedItems(prev => prev + 4);
+  };
+
+  const itemsToDisplay = items.slice(0, displayedItems);
+
   return (
     <>
       <div>
-        <select id="filter-items" defaultValue="">
+        <select id="filter-items" value={filter} onChange={onFilterChange}>
           <option value="">Default</option>
           <option value="price_low_to_high">Price, Low to High</option>
           <option value="price_high_to_low">Price, High to Low</option>
           <option value="likes_high_to_low">Most liked</option>
         </select>
       </div>
-      {new Array(8).fill(0).map((_, index) => (
-        <div
-          key={index}
-          className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
-          style={{ display: "block", backgroundSize: "cover" }}
-        >
-          <div className="nft__item">
-            <div className="author_list_pp">
-              <Link
-                to="/author"
-                data-bs-toggle="tooltip"
-                data-bs-placement="top"
-              >
-                <img className="lazy" src={AuthorImage} alt="" />
-                <i className="fa fa-check"></i>
-              </Link>
-            </div>
-            <div className="de_countdown">5h 30m 32s</div>
-
-            <div className="nft__item_wrap">
-              <div className="nft__item_extra">
-                <div className="nft__item_buttons">
-                  <button>Buy Now</button>
-                  <div className="nft__item_share">
-                    <h4>Share</h4>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-facebook fa-lg"></i>
-                    </a>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-twitter fa-lg"></i>
-                    </a>
-                    <a href="">
-                      <i className="fa fa-envelope fa-lg"></i>
-                    </a>
-                  </div>
-                </div>
+      {loading ? (
+        new Array(8).fill(0).map((_, index) => (
+          <div
+            key={index}
+            className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
+            style={{ display: "block", backgroundSize: "cover" }}
+          >
+            <div className="nft__item">
+              <div className="author_list_pp">
+                <div className="skeleton skeleton-avatar" style={{ width: "50px", height: "50px", borderRadius: "50%" }}></div>
               </div>
-              <Link to="/item-details">
-                <img src={nftImage} className="lazy nft__item_preview" alt="" />
-              </Link>
-            </div>
-            <div className="nft__item_info">
-              <Link to="/item-details">
-                <h4>Pinky Ocean</h4>
-              </Link>
-              <div className="nft__item_price">1.74 ETH</div>
-              <div className="nft__item_like">
-                <i className="fa fa-heart"></i>
-                <span>69</span>
+              <div className="de_countdown"></div>
+
+              <div className="nft__item_wrap">
+                <div className="skeleton skeleton-img" style={{ width: "100%", height: "300px" }}></div>
+              </div>
+              <div className="nft__item_info">
+                <div className="skeleton skeleton-title" style={{ width: "70%", marginBottom: "10px" }}></div>
+                <div className="skeleton skeleton-text" style={{ width: "40%" }}></div>
               </div>
             </div>
           </div>
+        ))
+      ) : (
+        itemsToDisplay.map((item, index) => (
+          <NFTCard key={index} item={item} variant="grid" />
+        ))
+      )}
+      {!loading && itemsToDisplay.length < items.length && (
+        <div className="col-md-12 text-center">
+          <button onClick={handleLoadMore} id="loadmore" className="btn-main lead">
+            Load more
+          </button>
         </div>
-      ))}
-      <div className="col-md-12 text-center">
-        <Link to="" id="loadmore" className="btn-main lead">
-          Load more
-        </Link>
-      </div>
+      )}
     </>
   );
 };
